consolidated Codalot
