--- test/codalot.cpp
@@ -1,8 +1,0 @@
-#include "knight_position.h"
-
-class Codalot {
-public:
-    virtual void setKnight(int id, KnightPosition position) = 0;
-    virtual void process() = 0;
-    virtual int calculateEarnedXp() = 0;
-};
