--- test/knight_position.cpp
@@ -1,4 +1,0 @@
-typedef enum KnightPosition_t {
-    TAVERN,
-    TRAINING_YARD
-} KnightPosition;
